Updates sparkscan-client to 0.1.0 and adds middleware support
Updates the sparkscan-client crate to version 0.1.0.

Adds support for reqwest-middleware via a feature flag.
This allows users to add middleware to the reqwest client for
features such as retry policies or logging.

Adds cfg-if dependency to handle conditional compilation based on the
"middleware" feature.

Moves reqwest-middleware and cfg-if dependencies to the workspace
level for better dependency management.
Adds MPL-2.0 License file.