feat(sparkscan-ws): adds sparkscan-ws crate (#3)
* feat(sparkscan-ws): adds sparkscan-ws crate

Introduces a new crate, sparkscan-ws, providing a WebSocket client for the SparkScan API.

This client includes:
- Asynchronous communication via tokio-centrifuge.
- Type-safe message handling with serde.
- Topic-based subscriptions.
- Automatic code generation from schemas.

* Removes custom topic subscription

Removes the ability to subscribe to custom topics.

This change enforces the use of predefined, strongly-typed topics,
improving code maintainability and reducing the risk of unexpected data formats.
It simplifies the API surface by removing the `subscribe_to_custom` method and
related logic.

* Refactors WebSocket client example and API

Improves the clarity and organization of the WebSocket
client example, demonstrating various subscription patterns
and message handling.

Updates the API documentation and descriptions to better
reflect the client's purpose and usage.

- Enhances example descriptions
- Clarifies connection management
- Improves subscription creation workflows

* feat(sparkscan-ws): adds configuration for sparkscan-ws crate

Adds a new package configuration for the sparkscan-ws crate to manage its release process, including changelog generation and git tagging.