chore(sparkscan): fmt
Adds a custom deserializer attribute to fields that require
deserialization of i128 or Option types within untagged enums,
but only if the attribute is not already present. This ensures correct
handling of these types during deserialization.

Also, fixes a bug where it was not unwrapping the inner type path
correctly when checking for Option.